Move DOS county updates into own module.
